move kotlinx test to kotlinx-io module
