+++ fleeksoft-io/build.gradle.kts
@@ -35,41 +35,3 @@ mavenPublishing {
         }
     }
 }
-
-
-val rootPath = "generated/kotlin"
-kotlin {
-    sourceSets {
-        commonTest {
-            this.kotlin.srcDir(layout.buildDirectory.file(rootPath))
-        }
-    }
-}
-
-val generateBuildConfigFile: Task by tasks.creating {
-    group = "build setup"
-    val file = layout.buildDirectory.file("$rootPath/BuildConfig.kt")
-    outputs.file(file)
-
-    doLast {
-        val content =
-            """
-            package com.fleeksoft.io
-
-            object BuildConfig {
-                const val PROJECT_ROOT: String = "${rootProject.rootDir.absolutePath.replace("\\", "\\\\")}"
-            }
-            """.trimIndent()
-        file.get().asFile.writeText(content)
-    }
-}
-
-// Configure the generateBuildConfigFile task to run only for test tasks
-tasks.withType<Test>().configureEach {
-    dependsOn(generateBuildConfigFile)
-}
-
-// Ensure generateBuildConfigFile runs before compileKotlinJvm
-tasks.named("compileTestKotlinJvm").configure {
-    dependsOn(generateBuildConfigFile)
-}
