+++ fleeksoft-io/module.yaml
@@ -17,4 +17,3 @@ dependencies:
 
 test-dependencies:
   - $libs.kotlinx.io
-  - ../kotlinx-io
